update rgb2hex (#379)
rgb2hex is subject to a RegExp-based Denial of Service vulnerability in
versions prior to 0.1.6. Update package-lock.json so `npm ci` and
friends install a safe version. No idea if electron is vulnerable to
anything remotely resembling a realistic attack based on this deep
dependency, but why bother looking into it when you can just update and
move on with life?

Refs: https://snyk.io/vuln/npm:rgb2hex:20180429